+++ packages/sanity/src/core/bundles/util/const.ts
@@ -2,9 +2,6 @@
 
 import {type BundleDocument} from '../../store/bundles/types'
 
-/**
- * @internal
- */
 export const LATEST: Partial<BundleDocument> = {
   name: 'drafts',
   title: 'Latest',
