+++ packages/sanity/src/core/bundles/util/dummyGetters.ts
@@ -7,7 +7,6 @@ import {type BundleDocument} from '../../store/bundles/types'
 
 /**
  * Returns all versions of a document
- * @internal
  *
  * @param documentId - document id
  * @param client - sanity client
