+++ packages/sanity/src/core/store/bundles/__workshop__/BundlesStoreStory.tsx
@@ -2,9 +2,9 @@ import {Card, Flex, Stack, Text} from '@sanity/ui'
 import {type ComponentType, type FormEvent, useCallback, useState} from 'react'
 
 import {Button} from '../../../../ui-components'
+import {BundleForm} from '../../../bundles/components/dialog/BundleForm'
 import {LoadingBlock} from '../../../components/loadingBlock/LoadingBlock'
 import {AddonDatasetProvider} from '../../../studio/addonDataset/AddonDatasetProvider'
-import {BundleForm} from '../../../versions/components/dialog/BundleForm'
 import {BundlesProvider, useBundles} from '../BundlesProvider'
 import {type BundleDocument} from '../types'
 import {useBundleOperations} from '../useBundleOperations'
