+++ packages/sanity/src/core/studio/components/navbar/StudioNavbar.tsx
@@ -19,7 +19,6 @@ import {Button, TooltipDelayGroupProvider} from '../../../../ui-components'
 import {type NavbarProps} from '../../../config/studio/types'
 import {isDev} from '../../../environment'
 import {useTranslation} from '../../../i18n'
-import {GlobalBundleMenu} from '../../../versions/components/GlobalBundleMenu'
 import {useToolMenuComponent} from '../../studio-components-hooks'
 import {useWorkspace} from '../../workspace'
 import {ConfigIssuesButton} from './configIssues/ConfigIssuesButton'
@@ -216,7 +215,7 @@ export function StudioNavbar(props: Omit<NavbarProps, 'renderDefault'>) {
                 </Flex>
                 {/* Versions button */}
                 <Flex gap={2}>
-                  <GlobalBundleMenu />
+                  <GlobalPerspectiveMenu />
                   {/* New document button */}
                   <NewDocumentButton
                     {...newDocumentOptions}
