--- packages/sanity/src/core/versions/components/GlobalBundleMenu.tsx
@@ -1,59 +1,0 @@
-import {AddIcon} from '@sanity/icons'
-import {Button, MenuItem} from '@sanity/ui'
-import {useCallback, useContext, useState} from 'react'
-
-import {
-  VersionContext,
-  type VersionContextValue,
-} from '../../../_singletons/core/form/VersionContext'
-import {useBundles} from '../../store/bundles/BundlesProvider'
-import {LATEST} from '../util/const'
-import {BundleBadge} from './BundleBadge'
-import {BundleMenu} from './BundleMenu'
-import {CreateBundleDialog} from './dialog/CreateBundleDialog'
-
-export function GlobalBundleMenu(): JSX.Element {
-  const {data: bundles, loading} = useBundles()
-
-  const [createBundleDialogOpen, setCreateBundleDialogOpen] = useState(false)
-
-  // TODO MAKE SURE THIS IS HOW WE WANT TO DO THIS
-  const {currentVersion, isDraft} = useContext<VersionContextValue>(VersionContext)
-
-  /* create new bundle */
-  const handleCreateBundleClick = useCallback(() => {
-    setCreateBundleDialogOpen(true)
-  }, [])
-
-  const handleClose = useCallback(() => {
-    setCreateBundleDialogOpen(false)
-  }, [])
-
-  return (
-    <>
-      <BundleMenu
-        button={
-          <Button mode="bleed" padding={0} radius="full">
-            <BundleBadge
-              hue={currentVersion?.hue}
-              icon={isDraft ? undefined : currentVersion?.icon}
-              openButton
-              padding={2}
-              title={isDraft ? LATEST.title : currentVersion?.title}
-            />
-          </Button>
-        }
-        bundles={bundles}
-        loading={loading}
-        actions={
-          // localize text
-          <MenuItem icon={AddIcon} onClick={handleCreateBundleClick} text="Create release" />
-        }
-      />
-
-      {createBundleDialogOpen && (
-        <CreateBundleDialog onCancel={handleClose} onCreate={handleClose} />
-      )}
-    </>
-  )
-}
