--- packages/sanity/src/core/versions/util/const.ts
@@ -1,11 +1,0 @@
-/*  TEMPORARY  DUMMY DATA */
-
-import {type BundleDocument} from '../../store/bundles/types'
-
-export const LATEST: Partial<BundleDocument> = {
-  name: 'drafts',
-  title: 'Latest',
-  icon: undefined,
-  tone: 'gray',
-  publishAt: '',
-}
