+++ packages/sanity/src/structure/panes/document/documentPanel/header/DocumentHeaderTitle.tsx
@@ -3,9 +3,9 @@ import {Flex, Text} from '@sanity/ui'
 import {createElement, type ReactElement} from 'react'
 import {unstable_useValuePreview as useValuePreview, useTranslation} from 'sanity'
 
-import {DocumentVersionMenu} from '../../../../../core/versions/components/panes/DocumentVersionMenu'
 import {structureLocaleNamespace} from '../../../../i18n'
 import {useDocumentPane} from '../../useDocumentPane'
+import {DocumentPerspectiveMenu} from './perspective/DocumentPerspectiveMenu'
 
 export function DocumentHeaderTitle(): ReactElement {
   const {documentId, connectionState, schemaType, title, value: documentValue} = useDocumentPane()
@@ -59,7 +59,7 @@ export function DocumentHeaderTitle(): ReactElement {
       </Flex>
 
       <Flex flex="none" gap={1}>
-        <DocumentVersionMenu documentId={documentId} />
+        <DocumentPerspectiveMenu documentId={documentId} />
       </Flex>
     </Flex>
   )
