+++ packages/sanity/src/structure/panes/document/documentPanel/header/perspective/DocumentPerspectiveMenu.tsx
@@ -1,20 +1,15 @@
 import {ChevronDownIcon} from '@sanity/icons'
 import {Box, Button} from '@sanity/ui'
 import {useCallback, useEffect, useState} from 'react'
-import {
-  BundleBadge,
-  type BundleDocument,
-  BundleMenu,
-  DEFAULT_STUDIO_CLIENT_OPTIONS,
-  getAllVersionsOfDocument,
-  LATEST,
-  useClient,
-  usePerspective,
-} from 'sanity'
+import {DEFAULT_STUDIO_CLIENT_OPTIONS, useClient} from 'sanity'
 
-// FIXME
-// eslint-disable-next-line boundaries/element-types
-import {useBundles} from '../../../../../../core/store/bundles/useBundles'
+import {BundleBadge} from '../../../../../../core/bundles/components/BundleBadge'
+import {BundleMenu} from '../../../../../../core/bundles/components/BundleMenu'
+import {usePerspective} from '../../../../../../core/bundles/hooks/usePerspective'
+import {LATEST} from '../../../../../../core/bundles/util/const'
+import {getAllVersionsOfDocument} from '../../../../../../core/bundles/util/dummyGetters'
+import {useBundles} from '../../../../../../core/store/bundles/BundlesProvider'
+import {type BundleDocument} from '../../../../../../core/store/bundles/types'
 
 export function DocumentPerspectiveMenu(props: {documentId: string}): JSX.Element {
   const {documentId} = props
